Fix: 인터페이스와 구현클래스 분리
